Rename component to TinyboxGallery
Closes https://github.com/kytta/vue-tinybox/issues/44
