+++ src/index.js
@@ -1,14 +1,13 @@
-import Component from "./tinybox.vue";
+import component from "./tinybox-gallery.vue";
 
 // Allow installing with app.use()
 // Source: https://github.com/houtaroy/vue-component-template/blob/619cacae5b000e4afdc623f2cb5f64fa522214ca/package/utils/ComponentUtil.ts
-const withInstall = (comp) => {
-	const c = comp;
-	c.install = function (app) {
-		app.component(c.name, Component);
+// IIFE approach inspired by vue-sfc-rollup
+export default ((_component) => {
+	const installable = _component;
+	installable.install = function (app) {
+		app.component("TinyboxGallery", installable);
 	};
 
-	return comp;
-};
-
-export default withInstall(Component);
+	return _component;
+})(component);
